fixed type

+++ lib/io/hardware_can.cpp
@@ -23,7 +23,7 @@ std::optional<std::shared_ptr<HardwareCan>> HardwareCan::create(
     close(socket_id);
     return std::nullopt;
   }
-  const sockaddr_can socket_address = {AF_CAN, interface_index};
+  const sockaddr_can socket_address = {AF_CAN, static_cast<int>(interface_index)};
   const int bind_status
     = bind(socket_id, reinterpret_cast<const sockaddr *>(&socket_address), sizeof(socket_address));
   if (bind_status < 0) {
